refactor: flows-test-kit mocks
(cherry picked from commit 13d3a3abcebef99ade81d642bcd5463bb109a305)
